+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.8.0 (2019-XX-XX)
 
 - Fixed `AIOHTTP` middleware to catch any possible exception, #6
+- Fixed `AIOHTTP` middleware to properly handle variable routes, #8
 
 ## 0.7.0 (2019-08-14)